requirements.txt
arxiv==2.1.3
groq>=1.30.0
openai>=1.30.0
httpx>=0.27.0
sentence-transformers>=2.7.0
networkx>=3.3
streamlit>=1.35.0
matplotlib>=3.9.0
numpy>=1.26.0
tqdm>=4.66.0
python-dotenv>=1.0.0
